[config] mempool_config own file
per title
