+++ config/src/config/mempool_config.rs
@@ -1,0 +1,37 @@
+// Copyright (c) The Libra Core Contributors
+// SPDX-License-Identifier: Apache-2.0
+
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Debug, Deserialize, Serialize)]
+#[serde(default)]
+pub struct MempoolConfig {
+    pub broadcast_transactions: bool,
+    pub shared_mempool_tick_interval_ms: u64,
+    pub shared_mempool_batch_size: usize,
+    pub shared_mempool_max_concurrent_inbound_syncs: usize,
+    pub capacity: usize,
+    // max number of transactions per user in Mempool
+    pub capacity_per_user: usize,
+    pub system_transaction_timeout_secs: u64,
+    pub system_transaction_gc_interval_ms: u64,
+    pub mempool_service_port: u16,
+    pub address: String,
+}
+
+impl Default for MempoolConfig {
+    fn default() -> MempoolConfig {
+        MempoolConfig {
+            broadcast_transactions: true,
+            shared_mempool_tick_interval_ms: 50,
+            shared_mempool_batch_size: 100,
+            shared_mempool_max_concurrent_inbound_syncs: 100,
+            capacity: 1_000_000,
+            capacity_per_user: 100,
+            system_transaction_timeout_secs: 86400,
+            address: "localhost".to_string(),
+            mempool_service_port: 6182,
+            system_transaction_gc_interval_ms: 180_000,
+        }
+    }
+}
